fix(#1456): keep OAuth server helpers out of the client bundle
`build:e2e` failed with "Rolldown failed to resolve import cloudflare:workers"
from device-auth.ts: the Start compiler strips server-fn handlers from the
client build but keeps plain exports, and `src/functions/middleware.ts` and
`src/functions/oauth-consent.ts` had started exporting helpers for their tests
(`resolveRequestPrincipal`, `loadOAuthConsentContext`,
`performAuthorizedAppRevoke`), which kept the Better Auth config reachable
from the browser bundle.

Move them into server-only modules — `src/lib/auth/request-principal.ts` and
`src/lib/auth/oauth-consent.ts` — referenced only from handler bodies, and
move their tests alongside. The client-server boundary test now passes again.
